requirements.txt
Django==4.0.1
django-cors-headers==3.10.1
PyMySQL==1.0.2
cryptography==36.0.1
pyjwt==2.3.0
requests==2.27.1
python-dotenv==0.19.2
gunicorn==20.1.0
pytz==2021.3
